fix: remove input tests in main component
