+++ src/components/Form/Form.test.tsx
@@ -1,51 +1,11 @@
-import { render, screen } from '@testing-library/react';
+import { fireEvent, render, screen } from '@testing-library/react';
 import React from 'react';
 import Form from './Form';
+import userEvent from '@testing-library/user-event';
 
 describe('Form', () => {
   it('render inputs', () => {
-    const inputTitleRef = React.createRef<HTMLInputElement>();
-    const inputDateRef = React.createRef<HTMLInputElement>();
-    const inputSelectRef = React.createRef<HTMLSelectElement>();
-    const inputPosterRef = React.createRef<HTMLInputElement>();
-    const inputRadioRumoredRef = React.createRef<HTMLInputElement>();
-    const inputRadioPlannedRef = React.createRef<HTMLInputElement>();
-    const inputRadioInProductionRef = React.createRef<HTMLInputElement>();
-    const inputRadioPostProductionRef = React.createRef<HTMLInputElement>();
-    const inputRadioReleasedRef = React.createRef<HTMLInputElement>();
-    const inputAdultRef = React.createRef<HTMLInputElement>();
-    const formRef = React.createRef<HTMLFormElement>();
-    render(
-      <Form
-        formRef={formRef}
-        onSubmit={() => {}}
-        inputTitleRef={inputTitleRef}
-        inputDateRef={inputDateRef}
-        inputSelectRef={inputSelectRef}
-        inputPosterRef={inputPosterRef}
-        inputRadioRumoredRef={inputRadioRumoredRef}
-        inputRadioPlannedRef={inputRadioPlannedRef}
-        inputRadioInProductionRef={inputRadioInProductionRef}
-        inputRadioPostProductionRef={inputRadioPostProductionRef}
-        inputRadioReleasedRef={inputRadioReleasedRef}
-        inputAdultRef={inputAdultRef}
-        cardState={{
-          title: null,
-          date: null,
-          ganre: null,
-          image: null,
-          status: null,
-          adult: false,
-          errors: {
-            invalidTitle: '',
-            invalidGanre: '',
-            invalidDate: '',
-            invalidPoster: '',
-            invalidStatus: '',
-          },
-        }}
-      />
-    );
+    render(<Form setFormData={() => {}} lengthArray={1}></Form>);
     const radioInputs = screen.getAllByRole('radio');
     const checkboxInputs = screen.getAllByRole('checkbox');
     expect(radioInputs).toHaveLength(5);
@@ -53,5 +13,13 @@ describe('Form', () => {
     expect(screen.getByLabelText('Title:')).toBeInTheDocument();
     expect(screen.getByLabelText('Select release date')).toBeInTheDocument();
     expect(screen.getByLabelText('Select ganre:')).toBeInTheDocument();
+
+    const selectElement = screen.getByRole('combobox') as HTMLSelectElement;
+    userEvent.selectOptions(selectElement, ['Action']);
+    expect(screen.getByRole('option', { name: 'Action' })).toBeInTheDocument();
+
+    const input = screen.getByPlaceholderText('movie title');
+    fireEvent.change(input, { target: { value: 'Godfather' } });
+    expect((input as HTMLInputElement).value).toBe('Godfather');
   });
 });
